+++ aop/modelButtonWidget.cpp
@@ -14,7 +14,7 @@ void ModelButtonWidget::doRenderInside() {
 	// init the persective matrix
 	glMatrixMode( GL_PROJECTION );
 	glPushMatrix();
-	glLoadMatrix( Eigen::createPerspectiveProjectionMatrix( 60.0, realSize[0] / realSize[1], 0.001, 100.0 ) );
+	glLoadMatrix( Eigen::createPerspectiveProjectionMatrix( 60.0f, realSize[0] / realSize[1], 0.001f, 100.0f ) );
 
 	glMatrixMode( GL_MODELVIEW );
 	glPushMatrix();
@@ -25,14 +25,14 @@ void ModelButtonWidget::doRenderInside() {
 
 	const float maxWidth = std::max( boundingBoxSize.x(), boundingBoxSize.z() );
 	const float maxDiagonal = sqrt( maxWidth * maxWidth + boundingBoxSize.y() * boundingBoxSize.y() );
-	const float scaleFactor = 2 * sqrt( 2.0 / 3.0 ) / (maxDiagonal * sqrt(2.0));
+	const float scaleFactor = 2 * sqrt( 2.0f / 3.0f ) / (maxDiagonal * sqrt(2.0f));
 
 	glLoadMatrix( Eigen::createLookAtMatrix( worldViewerPosition, Vector3f::Zero(), Vector3f::UnitY() ) );
 
 	// see note R3 for the computation
 	Eigen::glScale( Vector3f::Constant( scaleFactor ) );
 
-	float angle = time * 2 * Math::PI / 10.0;
+	float angle = static_cast< float >( time * 2 * Math::PI / 10.0f );
 	glMultMatrix( Affine3f(AngleAxisf( angle, Vector3f::UnitY() )).matrix() );
 
 	glTranslate( -boundingBox.center() );
@@ -43,7 +43,7 @@ void ModelButtonWidget::doRenderInside() {
 	glEnable( GL_DEPTH_TEST );
 
 	{
-		ViewportContext viewport( realMinCorner[0], realMinCorner[1], realSize[0], realSize[1] );
+		ViewportContext viewport( (int) realMinCorner[0], (int) realMinCorner[1], (int) realSize[0], (int) realSize[1] );
 
 		renderer.renderModel( worldViewerPosition, modelIndex );
 	}
